Introduce topic box and add sorting
The topic box makes it easy to add new topics.  The index view has
changed to only show the topics which have been saved.  They are also
sorted on the clientside now, to make displaying easier.  Sadly, this
doesn't mix well with the query support which we should be using.

The topic box allows us to create a new topic with a score.  This should
clearly be disallowed, as that'll be the responsability of the
plus-one-service.  However, this approach is easier for debugging.