Add draft plus-one-button component
The component isn't as pretty as it should be yet, but it should be
sufficient for our current efforts of voting :)

There's a special issue here, in that a page reload will assume that you
haven't voted on anything.  We either need to build a custom
microservice to return these results, or we have to adapt
mu-cl-resources with computed results.